autonn/autonn/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.25
django-cors-headers==3.11.0
djangorestframework==3.15.0
djangorestframework-jwt==1.11.0
easydict==1.10
idna==3.3
matplotlib>=3.5.1
numpy==1.22.3
#ofa
opencv-python==4.7.0.72
pandas==1.4.2
Pillow>=8.4.0
psycopg2-binary==2.9.3
#pytorch-lightning==1.6.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
seaborn==0.11.2
setuptools==59.5.0
sqlparse==0.4.2
tensorboard
tensorboardX==2.6
thop>=0.1.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.62.3
typing_extensions==4.1.1
urllib3==1.26.7
#pyyaml==6.0
protobuf==3.20.1
onnx==1.12.0
onnxsim==0.4.33
